FEAT(json schema ir): Replace intermediate representation with JSON schema
The previous intermediate representation was hand-rolled, and new
typescript features or difficult ones (native enums, mapped types) had
poor support. For example, enums-as-keys to a mapped Record<Enum,
string> type failed to compile.

This commit replaces the intermediate representation with JSONSchema, a
known validation format. We use the 'ts-json-schema-generator' on the
frontend to generate JSONSchema based on TypeScript, and 'jsonschema' on
the backend to validate that the schema is correct.

Custom tags are supported via the "fullDescription" option, where the
entire JSDoc is put into the AST and we parse it in the backend. This is
heavyweight and will lead to larger validators, but
ts-json-schema-generator doesn't support arbitrary keys
(https://github.com/vega/ts-json-schema-generator/issues/2281)